server/requirements.txt
# FCC
Flask==2.3.2
Response
paho-mqtt

# viz system
opencv-python
numpy

# CDC
redis
pyyaml
openpyxl